+++ spec/components/schemas/content/facebook/all.ts
@@ -1,35 +1,12 @@
+import * as _ from 'lodash';
 import { SchemaObject } from 'openapi3-ts';
-import { ref as textRef } from '../text';
-import { ref as fileRef } from '../file';
-import { ref as quickRepliesRef } from '../replyable-text';
-import { ref as cardRef } from '../card';
-import { ref as carouselRef } from '../carousel';
 import { createComponentRef } from '../../../../../utils/ref';
+import { ref as optInResponseRef } from '../optin_response';
+import mt from './mt';
 
-const all: SchemaObject = {
-  title: 'Facebook',
-  oneOf: [{
-    $ref: textRef,
-  }, {
-    $ref: fileRef,
-  }, {
-    $ref: cardRef,
-  }, {
-    $ref: carouselRef,
-  }, {
-    $ref: quickRepliesRef,
-  }],
-  discriminator: {
-    propertyName: 'type',
-    mapping: {
-      text: textRef,
-      file: fileRef,
-      card: cardRef,
-      carousel: carouselRef,
-      replyable_text: quickRepliesRef,
-    },
-  },
-};
+const all: SchemaObject = _.cloneDeep(mt);
+all.oneOf.push({ $ref: optInResponseRef });
+all.discriminator.mapping.optin_response = optInResponseRef;
 
 export const ref = createComponentRef(__filename);
 export default all;
