+++ spec/components/schemas/message/facebook.ts
@@ -1,6 +1,6 @@
 import { SchemaObject } from 'openapi3-ts';
 import { ref as baseRef } from './base';
-import { ref as allContentsRef } from '../content/facebook/all';
+import { ref as allContentsRef } from '../content/facebook/mt';
 import { createComponentRef } from '../../../../utils/ref';
 
 const all: SchemaObject = {
